+++ SolarPowerPlant/Models/DTOs/SolarPlantDTOs/GetSolarPlantDTO.cs
@@ -9,7 +9,7 @@ namespace SolarPowerAPI.Models.DTOs.SolarPlantDTOs
 
         public string? Name { get; set; }
 
-        public int InstalledPower { get; set; }
+        public double InstalledPower { get; set; }
 
         public DateTime DateOfInstallation { get; set; }
