+++ SolarPowerPlant/Models/Entities/SolarPlant.cs
@@ -5,13 +5,14 @@ namespace SolarPowerAPI.Models.Entities
 {
     public class SolarPlant
     {
+        //TODO: Update database (create new migration)
         [Key]
         public Guid Id { get; set; }
 
         public string? Name { get; set; }
 
         [Required]
-        public int InstalledPower { get; set; }
+        public double InstalledPower { get; set; }
 
         [Required]
         public DateTime DateOfInstallation { get; set; }
